VuVanNghia20206205

+++ workspace/others/push/modules/MyFormat.py
@@ -66,7 +66,20 @@ class MyFormat:
     #         with open(file_markdown, 'w', encoding="utf-8") as file:
     #             file.write(contents)
 
-    def basic(path):
+    def basic1(path):
+        with open(path, 'r', encoding="utf-8") as file:
+            contents = file.read()
+
+
+        while '  ' in contents:
+            contents = contents.replace('  ', ' ')
+        contents = '\n'.join(line.strip() for line in contents.split('\n'))
+        while "\n\n\n" in contents:
+            contents = contents.replace("\n\n\n", "\n\n")
+        contents = contents.lstrip('\n')
+        with open(path, 'w', encoding="utf-8") as file:
+            file.write(contents)
+    def basic2(path):
         with open(path, 'r', encoding="utf-8") as file:
             contents = file.read()
 
@@ -81,31 +94,36 @@ class MyFormat:
         with open(path, 'w', encoding="utf-8") as file:
             file.write(contents)
 
-    def latex(contents):
-        latex_files = glob.glob(os.path.join(
-            contents, f'**/*.tex'), recursive=True)
+    def latex(contents): 
+        if os.path.isfile(contents):
+            MyFormat.basic1(contents) 
+        else:
 
-        for latex_file in latex_files:
-            with open(latex_file, 'r', encoding="utf-8") as file:
-                contents = file.read()
 
-            while '  ' in contents:
-                contents = contents.replace('  ', ' ')
+            latex_files = glob.glob(os.path.join(
+                contents, f'**/*.tex'), recursive=True)
 
-            while '( ' in contents:
-                contents = contents.replace('( ', '(')
-            while ' )' in contents:
-                contents = contents.replace(' )', ')')
-            while '[ ' in contents:
-                contents = contents.replace('[ ', '[')
-            while ' ]' in contents:
-                contents = contents.replace(' ]', ']')
-            while '{ ' in contents:
-                contents = contents.replace('{ ', '{')
-            while ' }' in contents:
-                contents = contents.replace(' }', '}')
+            for latex_file in latex_files:
+                with open(latex_file, 'r', encoding="utf-8") as file:
+                    contents = file.read()
 
-            with open(latex_file, 'w', encoding="utf-8") as file:
-                file.write(contents)
+                while '  ' in contents:
+                    contents = contents.replace('  ', ' ')
 
-            MyFormat.basic(latex_file)
+                while '( ' in contents:
+                    contents = contents.replace('( ', '(')
+                while ' )' in contents:
+                    contents = contents.replace(' )', ')')
+                while '[ ' in contents:
+                    contents = contents.replace('[ ', '[')
+                while ' ]' in contents:
+                    contents = contents.replace(' ]', ']')
+                while '{ ' in contents:
+                    contents = contents.replace('{ ', '{')
+                while ' }' in contents:
+                    contents = contents.replace(' }', '}')
+
+                with open(latex_file, 'w', encoding="utf-8") as file:
+                    file.write(contents)
+
+                MyFormat.basic2(latex_file)
